+++ src/main/java/com/nisovin/shopkeepers/ui/defaults/confirmations/ConfirmationUIConfig.java
@@ -1,0 +1,10 @@
+package com.nisovin.shopkeepers.ui.defaults.confirmations;
+
+import java.util.List;
+
+public interface ConfirmationUIConfig {
+
+	public String getTitle();
+
+	public List<String> getConfirmationLore();
+}