+++ src/main/java/com/nisovin/shopkeepers/ui/defaults/confirmations/ConfirmationUIType.java
@@ -1,0 +1,10 @@
+package com.nisovin.shopkeepers.ui.defaults.confirmations;
+
+import com.nisovin.shopkeepers.ui.AbstractUIType;
+
+public class ConfirmationUIType extends AbstractUIType {
+
+	public ConfirmationUIType() {
+		super("confirmation", null);
+	}
+}